darwin-workflow/requirements.txt
setuptools~=67.8.0
pydantic~=1.10.12
shortuuid~=1.0.11
requests~=2.28.1
PyYAML~=6.0.1
fastapi~=0.95.1
uvicorn[standard]~=0.22.0
typeguard~=2.13.3
ddtrace~=3.9.3
httpx~=0.23.3
pytz~=2023.3.post1
pytest-asyncio~=0.21.1